Move ServerSubscription to node package
... and update for new logic using new readers.
