+++ packages/protocol/src/interaction/index.ts
@@ -11,6 +11,5 @@ export * from "./EventDataDecoder.js";
 export * from "./InteractionClient.js";
 export * from "./InteractionEndpointStructure.js";
 export * from "./InteractionMessenger.js";
-export * from "./ServerSubscription.js";
 export * from "./Subscription.js";
 export * from "./SubscriptionClient.js";
